Restructure the library
